% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blinder.R
\name{handleClash}
\alias{handleClash}
\title{Appends underscore and increasing number to the end of a filename until there is no longer a clash}
\usage{
handleClash(base, ext, dir)
}
\arguments{
\item{base}{Initial file name, without extension}

\item{ext}{File extension}

\item{dir}{Directory path}
}
\description{
Appends underscore and increasing number to the end of a filename until there is no longer a clash
}
