% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blinder.R
\name{unblind}
\alias{unblind}
\title{Restore original file names.}
\usage{
unblind(target = NULL)
}
\arguments{
\item{target}{The directory containing blinded files to restore.  If this is not provided, the user will be prompted for a directory.}
}
\description{
Restore original file names.
}
